app/requirements.txt
django==3.2
django-extensions==3.1.5
django-debug-toolbar==3.4.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
split-settings==1.0.0
uWSGI==2.0.20